Refactored geolocation into its own file Optimized
